// doc-tools/lib/docauto/types.ts
export type TokenType =
  | 'text'
  | 'variable'
  | 'if'
  | 'elseif'
  | 'else'
  | 'endif'
  | 'list'
  | 'endlist'
  | 'template';

// Parsed Token
export interface Token {
  type: TokenType;
  content: string;
  variable?: string;
  condition?: string;
  formatters?: Formatter[];
  listVariable?: string;
  listFilters?: string;
  templateName?: string;
}

// Formatter
export interface Formatter {
  name: string;
  args?: string[];
}